rename webbapp to archivist, since that is the module name on the sight
