+++ api/adventureRecord.api.go
@@ -5,8 +5,8 @@ import (
 	"net/http"
 	"strconv"
 
-	"github.com/floodedrealms/adventure-archivist/internal/services"
-	"github.com/floodedrealms/adventure-archivist/types"
+	"github.com/floodedrealms/borderland-keep/internal/services"
+	"github.com/floodedrealms/borderland-keep/types"
 )
 
 type AdventureApi struct {
