+++ api/character.api.go
@@ -5,9 +5,9 @@ import (
 	"net/http"
 	"strconv"
 
-	"github.com/floodedrealms/adventure-archivist/internal/services"
-	"github.com/floodedrealms/adventure-archivist/internal/util"
-	"github.com/floodedrealms/adventure-archivist/types"
+	"github.com/floodedrealms/borderland-keep/internal/services"
+	"github.com/floodedrealms/borderland-keep/internal/util"
+	"github.com/floodedrealms/borderland-keep/types"
 )
 
 type CharacterApi struct {
