+++ archivist/home.pages.go
@@ -1,11 +1,11 @@
-package webapp
+package archivist
 
 import (
 	"net/http"
 	"strconv"
 
-	"github.com/floodedrealms/adventure-archivist/internal/services"
-	"github.com/floodedrealms/adventure-archivist/types"
+	"github.com/floodedrealms/borderland-keep/internal/services"
+	"github.com/floodedrealms/borderland-keep/types"
 )
 
 type HomePage struct {
