+++ archivist/pagedata.go
@@ -1,4 +1,4 @@
-package webapp
+package archivist
 
 type PageData interface {
 	IsAuthenticated() bool
