+++ archivist/path.go
@@ -1,4 +1,4 @@
-package webapp
+package archivist
 
 type path struct {
 	Display string
