+++ archivist/renderer.go
@@ -1,4 +1,4 @@
-package webapp
+package archivist
 
 import (
 	"bytes"
@@ -27,9 +27,9 @@ func NewRenderer() *Renderer {
 
 func (r *Renderer) mustLoadTemplates() {
 	wd, _ := os.Getwd()
-	pageDir := filepath.Join(wd, "/internal/templates/pages/*.html")
-	partialsDir := filepath.Join(wd, "/internal/templates/partials/*.html")
-	partialEditorsDir := filepath.Join(wd, "/internal/templates/partials/editors/*.html")
+	pageDir := filepath.Join(wd, "/archivist/templates/pages/*.html")
+	partialsDir := filepath.Join(wd, "/archivist/templates/partials/*.html")
+	partialEditorsDir := filepath.Join(wd, "/archivist/templates/partials/editors/*.html")
 
 	r.pageTemplates = template.Must(template.ParseGlob(pageDir))
 	r.partialTemplates = template.Must(template.ParseGlob(partialsDir))
