+++ archivist/templates/pages/adventurePage.html
