+++ archivist/templates/pages/index.html
@@ -24,7 +24,7 @@
                     centralization of ttrpg hobby discusions into walled gardens like Discord, Twitter or Reddit. This
                     site
                     aims to maintain serveral services, including: <a href="/tavern">A mastodon instance</a>, <a
-                        href="/guild"> a matrix homeserver </a>, and (eventually) <a href="/crier">a campaign
+                        href="/guild"> a matrix homeserver </a>, and (eventually) <a href="/archivist">a campaign
                         logging service.</a> Please take a look around, and enjoy your stay.</p>
             </hgroup>
             <hgroup class="box" id="proj">
