+++ archivist/templates/pages/newAdventurePage.html
