+++ archivist/templates/partials/adventureOverview.html
