+++ archivist/templates/partials/characterDetailsTableRow.html
