+++ archivist/templates/partials/combatRow.html
