+++ archivist/templates/partials/editors/adventureDetailsEdit.html
