+++ archivist/templates/partials/editors/characterDetailsTableEdit.html
