+++ archivist/templates/partials/editors/coinsEdit.html
