+++ archivist/templates/partials/editors/combatTableEdit.html
