+++ archivist/templates/partials/editors/lootTableEdit.html
