+++ archivist/templates/partials/editors/newAdventureForm.html
