+++ archivist/templates/partials/goldInputTD.html
