+++ archivist/templates/partials/lootFormRow.html
