+++ archivist/templates/partials/lootList.html
