+++ archivist/templates/partials/magicItemList.html
