+++ archivist/templates/partials/toggleGoldButtonTD.html
