+++ internal/repository/jsonTestDatabase.go
@@ -4,8 +4,8 @@ import (
 	"database/sql"
 	"errors"
 
-	"github.com/floodedrealms/adventure-archivist/internal/util"
-	"github.com/floodedrealms/adventure-archivist/types"
+	"github.com/floodedrealms/borderland-keep/internal/util"
+	"github.com/floodedrealms/borderland-keep/types"
 )
 
 type JSONRepo struct {
