+++ internal/services/campaignAction.service.go
@@ -1,8 +1,8 @@
 package services
 
 import (
-	"github.com/floodedrealms/adventure-archivist/internal/repository"
-	"github.com/floodedrealms/adventure-archivist/types"
+	"github.com/floodedrealms/borderland-keep/internal/repository"
+	"github.com/floodedrealms/borderland-keep/types"
 )
 
 type CampaignActionService struct {
